requirement.txt
python-telegram-bot==20.3
opencv-python
torchvision
facenet-pytorch
scipy